server/pyproject.toml
[project]
name = "earthsync-server"
version = "0.1.1"
description = "EarthSync FastAPI backend — Schumann Resonance processing server"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135",
    "uvicorn[standard]>=0.42",
    "asyncpg>=0.31",
    "redis[hiredis]>=7.4",
    "numpy>=2.4",
    "scipy>=1.17",
    "lmfit>=1.3",
    "pyjwt>=2.12",
    "pwdlib[argon2]>=0.3",
    "cryptography>=46",
    "pydantic-settings>=2.13",
    "fastapi-limiter>=0.2",
    "structlog>=25.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-cov>=7.1",
    "httpx>=0.28",
    "fakeredis[json]>=2.25",
    "ruff>=0.15",
    "pyright>=1.1",
    "bandit>=1.8",
    "vulture>=2.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/earthsync_server"]
